+++ docs/changelog.rst
@@ -18,6 +18,7 @@ v4.0.1
 ------
 
 - FIX: Broken docs on Sphinx 9 and due to changed Xopt API.
+- FIX: Tests broken by Xopt 2.7.
 
 v4.0.0
 ------